tests/wpt/web-platform-tests/workers/semantics/run-a-worker/001.html

<!--
postMessage(this === self);

/*
-->
<!doctype html>
<title>worker global scope, dedicated worker</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
async_test(function() {
  var worker = new Worker('#');
  worker.onmessage = this.step_func(function(e) {
    assert_true(e.data);
    this.done();
  });
});
</script>
<!--
*/
//-->
